feat(api): move /admin/users/{id}/anonymize route in src
