+++ api/lib/application/users/index.js
@@ -64,29 +64,6 @@ const register = async function (server) {
     },
     {
       method: 'POST',
-      path: '/api/admin/users/{id}/anonymize',
-      config: {
-        validate: {
-          params: Joi.object({
-            id: identifiersType.userId,
-          }),
-        },
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-              ])(request, h),
-          },
-        ],
-        handler: userController.anonymizeUser,
-        notes: ["- Permet à un administrateur d'anonymiser un utilisateur"],
-        tags: ['api', 'admin', 'user'],
-      },
-    },
-    {
-      method: 'POST',
       path: '/api/admin/users/{id}/add-pix-authentication-method',
       config: {
         pre: [
